requirements.txt
### Core Application ###
fastapi==0.109.1
uvicorn==0.27.1

### Database ###
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1  # Nouveau - pour les migrations
asyncpg==0.29.0  # Optionnel pour async

### Data Validation ###
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1  # Nouveau - validation d'emails

### Environment ###
python-dotenv==1.0.1

### Security ###
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Nouveau - pour le hachage sécurisé

### Development ###
pytest==8.1.1
httpx==0.27.0
pytest-asyncio==0.23.6
pytest-cov==4.1.0  # Nouveau - couverture de tests
ipdb==0.13.13  # Nouveau - debugger amélioré

### Documentation ###
mkdocs==1.5.3  # Nouveau - documentation
mkdocs-material==9.5.3  # Nouveau - thème doc






